requirements.txt
streamlit
opencv-python
numpy
Pillow
pillow-heif
scikit-image
